requirements.txt
torch
torchvision
opencv-python-headless
pillow
ultralytics
